// workspaces/marketplace/plugins/marketplace-common/src/oci/types.ts
/*
 * Copyright 2025 The Backstage Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export interface ImageRef {
  registry: string;
  repository: string;
  tag: string;
}

export interface TagsList {
  name: string;
  tags: string[];
}

export interface Manifest {
  tag: string;
  name: string;
  architecture: string;
  schemaVersion: number;
  history: Array<{ v1Compatibility: string }>;
  fsLayers: Array<{ blobSum: string }>;
}

export interface InspectResponse {
  name: string;
  digest: string;
  // repoTags: string[];
  created: string;
  architecture: string;
  labels: Record<string, string>;
  // size: number;
  // os: string;
  // layers: string[];
  // latersData: Array<{  }>;
  env: string[];
}